+++ redis/src/main/scala/zio/redis/RedisClusterConfig.scala
@@ -1,4 +1,4 @@
-package zio.redis.executor.cluster
+package zio.redis
 
 import zio.Chunk
