+++ redis/src/test/scala/zio/redis/KeysSpec.scala
@@ -2,8 +2,6 @@ package zio.redis
 
 import zio._
 import zio.redis.RedisError.ProtocolError
-import zio.redis.executor.RedisConnectionLive
-import zio.redis.executor.node.RedisNodeExecutorLive
 import zio.schema.codec.{Codec, ProtobufCodec}
 import zio.test.Assertion.{exists => _, _}
 import zio.test.TestAspect.{restore => _, _}
