+++ redis/src/test/scala/zio/redis/RedisClusterExecutorLiveSpec.scala
@@ -1,8 +1,6 @@
-package zio.redis.executor.cluster
+package zio.redis
 
-import zio.redis._
 import zio.redis.codec.CRC16
-import zio.redis.executor.RedisExecutor
 import zio.redis.options.Cluster.{Slot, SlotsAmount}
 import zio.test._
 import zio.{Chunk, Layer, ZIO, ZLayer, durationInt}
